Rename pathfinding to legacy pathfinding, make this pathfinding standard
