+++ src/monmove.cpp
@@ -36,8 +36,8 @@
 #include "mtype.h"
 #include "npc.h"
 #include "options.h"
+#include "legacy_pathfinding.h"
 #include "pathfinding.h"
-#include "pathfinding_dijikstra.h"
 #include "pimpl.h"
 #include "player.h"
 #include "point.h"
