+++ src/mtype.h
@@ -17,8 +17,8 @@
 #include "enum_bitset.h"
 #include "enums.h"
 #include "mattack_common.h"
+#include "legacy_pathfinding.h"
 #include "pathfinding.h"
-#include "pathfinding_dijikstra.h"
 #include "translations.h"
 #include "type_id.h"
 #include "units.h"
